# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de prueba NITSNETS",
        "description": "Esta es una primera versión de la API",
        "contact": "jorgepastor9120@gmail.com",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/v1/courts": {
            "get": {
                "tags": [
                    "Courts"
                ],
                "summary": "Mostrar el listado de pistas",
                "description": "Mostramos el listado de los regitros solicitados.",
                "operationId": "169d48b766db492f3b88a2dc18ac610a",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No se ha podido cargar el listado."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Courts"
                ],
                "summary": "Registra una pista",
                "description": "Registrar una pista.",
                "operationId": "8844919b2ce51a688141af6b7fd643d9",
                "requestBody": {
                    "description": "Registra una pista",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "sport_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Prueba"
                                    },
                                    "sport_id": {
                                        "type": "integer",
                                        "format": "int20",
                                        "example": "4"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "201": {
                        "description": "La pista ha sido creado"
                    },
                    "404": {
                        "description": "No se ha podido registrar la pista."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/courts/{id}": {
            "put": {
                "tags": [
                    "Courts"
                ],
                "summary": "Actualiza una pista",
                "description": "Actualizar un registro.",
                "operationId": "18e011c79872b3d2425194ac89f4f80b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id court",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int20"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Actualiza una pista",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "sport_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Prueba"
                                    },
                                    "sport_id": {
                                        "type": "integer",
                                        "format": "int20",
                                        "example": "4"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No se ha podido actualizar el registro."
                    },
                    "405": {
                        "description": "No se ha permitido actualizar el registro."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Courts"
                ],
                "summary": "Elimina una pista",
                "description": "Elimina un registro.",
                "operationId": "9e9bdad4534f30a493414cbfaea01429",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id court",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int20"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Elimina una pista"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No se ha podido eliminar el registro."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/courts/list": {
            "get": {
                "tags": [
                    "Courts"
                ],
                "summary": "Mostrar el listado de pistas libres según una fecha, un deporte y un miembro",
                "description": "Mostramos el listado de pistas disponibles según varios parametros.",
                "operationId": "cd9ca9af2216fa87e3886dd23de21f6e",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2023-10-25"
                    },
                    {
                        "name": "sport_id",
                        "in": "query",
                        "description": "sport_id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int20"
                        },
                        "example": "3"
                    },
                    {
                        "name": "member_id",
                        "in": "query",
                        "description": "member_id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int20"
                        },
                        "example": "30"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No se ha podido cargar el listado."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "http",
                "name": "authorization",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8000/api/v1/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}